+++ tests/features/token_cache/test_json_token_file_cache.py
@@ -1,4 +1,6 @@
 import datetime
+import logging
+import pathlib
 
 import pytest
 import jwt
@@ -54,7 +56,9 @@ def test_save_bearer_tokens(token_cache, tmp_path):
     assert same_cache.get_token("key2") == token2
 
 
-def test_save_bearer_token_exception_handling(token_cache, tmp_path, monkeypatch):
+def test_save_bearer_token_exception_handling(
+    token_cache, tmp_path, monkeypatch, caplog
+):
     def failing_dump(*args):
         raise Exception("Failure")
 
@@ -65,6 +69,8 @@ def test_save_bearer_token_exception_handling(token_cache, tmp_path, monkeypatch
     ) + datetime.timedelta(hours=1)
     token1 = jwt.encode({"exp": expiry_in_1_hour}, "secret")
 
+    caplog.set_level(logging.DEBUG)
+
     # Assert that the exception is not thrown
     token_cache._add_bearer_token("key1", token1)
 
@@ -73,10 +79,43 @@ def test_save_bearer_token_exception_handling(token_cache, tmp_path, monkeypatch
         same_cache.get_token("key1")
     assert str(exception_info.value) == "User was not authenticated."
 
+    assert caplog.messages == [
+        "Cannot save tokens.",
+        f'Inserting token expiring on {expiry_in_1_hour:%Y-%m-%d %H:%M:%S+00:00} with "key1" key.',
+        "Cannot load tokens.",
+        'Retrieving token with "key1" key.',
+        "Token cannot be found in cache.",
+        "User was not authenticated: key key1 cannot be found in [].",
+    ]
 
-def test_missing_token(token_cache):
+
+def test_missing_token_on_empty_cache(token_cache, caplog):
+    caplog.set_level(logging.DEBUG)
     with pytest.raises(httpx_auth.AuthenticationFailed):
         token_cache.get_token("key1")
+    assert caplog.messages == [
+        'Retrieving token with "key1" key.',
+        "No token loaded. Token cache does not exists.",
+        "Token cannot be found in cache.",
+        "User was not authenticated: key key1 cannot be found in [].",
+    ]
+
+
+def test_missing_token_on_non_empty_cache(token_cache, caplog):
+    expiry_in_1_hour = datetime.datetime.now(
+        datetime.timezone.utc
+    ) + datetime.timedelta(hours=1)
+    token1 = jwt.encode({"exp": expiry_in_1_hour}, "secret")
+    token_cache._add_bearer_token("key0", token1)
+
+    caplog.set_level(logging.DEBUG)
+    with pytest.raises(httpx_auth.AuthenticationFailed):
+        token_cache.get_token("key1")
+    assert caplog.messages == [
+        'Retrieving token with "key1" key.',
+        "Token cannot be found in cache.",
+        "User was not authenticated: key key1 cannot be found in ['key0'].",
+    ]
 
 
 def test_missing_token_function(token_cache):
@@ -105,3 +144,16 @@ def test_token_without_refresh_token(token_cache):
     # try to retrieve it
     retrieved_token = token_cache.get_token("key1")
     assert token == retrieved_token
+
+
+def test_unable_to_remove_cache(token_cache, tmp_path, monkeypatch, caplog):
+    def unlink_failure(*args):
+        raise PermissionError("You can create but can't delete")
+
+    monkeypatch.setattr(pathlib.Path, "unlink", unlink_failure)
+
+    caplog.set_level(logging.DEBUG)
+    # Assert that the exception is not thrown
+    token_cache.clear()
+
+    assert caplog.messages == ["Clearing token cache.", "Cannot remove tokens file."]